move usecase layer
