+++ src/index.ts
@@ -4,4 +4,5 @@
 export { attribute } from './attribute/index.js';
 export * from './model/index.js';
 export * from './parser/index.js';
-export * from './usecase/index.js';
+export * from './create-root-graph.js';
+export * from './to-dot.js';
