--- src/usecase/index.ts
@@ -1,4 +1,0 @@
-/**
- * @module ts-graphviz/usecase
- */
-export * from './builder.js';
