--- src/usecase/render/index.ts
@@ -1,1 +1,0 @@
-export * from './to-dot.js';
